+++ clients/groundstation.d.ts
@@ -36,6 +36,14 @@ declare class GroundStation extends Service {
    */
   createDataflowEndpointGroup(callback?: (err: AWSError, data: GroundStation.Types.DataflowEndpointGroupIdResponse) => void): Request<GroundStation.Types.DataflowEndpointGroupIdResponse, AWSError>;
   /**
+   * Creates an Ephemeris with the specified EphemerisData.
+   */
+  createEphemeris(params: GroundStation.Types.CreateEphemerisRequest, callback?: (err: AWSError, data: GroundStation.Types.EphemerisIdResponse) => void): Request<GroundStation.Types.EphemerisIdResponse, AWSError>;
+  /**
+   * Creates an Ephemeris with the specified EphemerisData.
+   */
+  createEphemeris(callback?: (err: AWSError, data: GroundStation.Types.EphemerisIdResponse) => void): Request<GroundStation.Types.EphemerisIdResponse, AWSError>;
+  /**
    * Creates a mission profile.  dataflowEdges is a list of lists of strings. Each lower level list of strings has two elements: a from ARN and a to ARN.
    */
   createMissionProfile(params: GroundStation.Types.CreateMissionProfileRequest, callback?: (err: AWSError, data: GroundStation.Types.MissionProfileIdResponse) => void): Request<GroundStation.Types.MissionProfileIdResponse, AWSError>;
@@ -60,6 +68,14 @@ declare class GroundStation extends Service {
    */
   deleteDataflowEndpointGroup(callback?: (err: AWSError, data: GroundStation.Types.DataflowEndpointGroupIdResponse) => void): Request<GroundStation.Types.DataflowEndpointGroupIdResponse, AWSError>;
   /**
+   * Deletes an ephemeris
+   */
+  deleteEphemeris(params: GroundStation.Types.DeleteEphemerisRequest, callback?: (err: AWSError, data: GroundStation.Types.EphemerisIdResponse) => void): Request<GroundStation.Types.EphemerisIdResponse, AWSError>;
+  /**
+   * Deletes an ephemeris
+   */
+  deleteEphemeris(callback?: (err: AWSError, data: GroundStation.Types.EphemerisIdResponse) => void): Request<GroundStation.Types.EphemerisIdResponse, AWSError>;
+  /**
    * Deletes a mission profile.
    */
   deleteMissionProfile(params: GroundStation.Types.DeleteMissionProfileRequest, callback?: (err: AWSError, data: GroundStation.Types.MissionProfileIdResponse) => void): Request<GroundStation.Types.MissionProfileIdResponse, AWSError>;
@@ -75,6 +91,14 @@ declare class GroundStation extends Service {
    * Describes an existing contact.
    */
   describeContact(callback?: (err: AWSError, data: GroundStation.Types.DescribeContactResponse) => void): Request<GroundStation.Types.DescribeContactResponse, AWSError>;
+  /**
+   * Describes an existing ephemeris.
+   */
+  describeEphemeris(params: GroundStation.Types.DescribeEphemerisRequest, callback?: (err: AWSError, data: GroundStation.Types.DescribeEphemerisResponse) => void): Request<GroundStation.Types.DescribeEphemerisResponse, AWSError>;
+  /**
+   * Describes an existing ephemeris.
+   */
+  describeEphemeris(callback?: (err: AWSError, data: GroundStation.Types.DescribeEphemerisResponse) => void): Request<GroundStation.Types.DescribeEphemerisResponse, AWSError>;
   /**
    * Returns Config information. Only one Config response can be returned.
    */
@@ -140,6 +164,14 @@ declare class GroundStation extends Service {
    */
   listDataflowEndpointGroups(callback?: (err: AWSError, data: GroundStation.Types.ListDataflowEndpointGroupsResponse) => void): Request<GroundStation.Types.ListDataflowEndpointGroupsResponse, AWSError>;
   /**
+   * List existing ephemerides.
+   */
+  listEphemerides(params: GroundStation.Types.ListEphemeridesRequest, callback?: (err: AWSError, data: GroundStation.Types.ListEphemeridesResponse) => void): Request<GroundStation.Types.ListEphemeridesResponse, AWSError>;
+  /**
+   * List existing ephemerides.
+   */
+  listEphemerides(callback?: (err: AWSError, data: GroundStation.Types.ListEphemeridesResponse) => void): Request<GroundStation.Types.ListEphemeridesResponse, AWSError>;
+  /**
    * Returns a list of ground stations. 
    */
   listGroundStations(params: GroundStation.Types.ListGroundStationsRequest, callback?: (err: AWSError, data: GroundStation.Types.ListGroundStationsResponse) => void): Request<GroundStation.Types.ListGroundStationsResponse, AWSError>;
@@ -204,6 +236,14 @@ declare class GroundStation extends Service {
    */
   updateConfig(callback?: (err: AWSError, data: GroundStation.Types.ConfigIdResponse) => void): Request<GroundStation.Types.ConfigIdResponse, AWSError>;
   /**
+   * Updates an existing ephemeris
+   */
+  updateEphemeris(params: GroundStation.Types.UpdateEphemerisRequest, callback?: (err: AWSError, data: GroundStation.Types.EphemerisIdResponse) => void): Request<GroundStation.Types.EphemerisIdResponse, AWSError>;
+  /**
+   * Updates an existing ephemeris
+   */
+  updateEphemeris(callback?: (err: AWSError, data: GroundStation.Types.EphemerisIdResponse) => void): Request<GroundStation.Types.EphemerisIdResponse, AWSError>;
+  /**
    * Updates a mission profile. Updating a mission profile will not update the execution parameters for existing future contacts.
    */
   updateMissionProfile(params: GroundStation.Types.UpdateMissionProfileRequest, callback?: (err: AWSError, data: GroundStation.Types.MissionProfileIdResponse) => void): Request<GroundStation.Types.MissionProfileIdResponse, AWSError>;
@@ -213,6 +253,7 @@ declare class GroundStation extends Service {
   updateMissionProfile(callback?: (err: AWSError, data: GroundStation.Types.MissionProfileIdResponse) => void): Request<GroundStation.Types.MissionProfileIdResponse, AWSError>;
 }
 declare namespace GroundStation {
+  export type AWSRegion = string;
   export type AngleUnits = "DEGREE_ANGLE"|"RADIAN"|string;
   export interface AntennaDemodDecodeDetails {
     /**
@@ -254,6 +295,7 @@ declare namespace GroundStation {
      */
     transmitDisabled?: Boolean;
   }
+  export type AnyArn = string;
   export type BandwidthUnits = "GHz"|"MHz"|"kHz"|string;
   export type Boolean = boolean;
   export type BucketArn = string;
@@ -261,7 +303,7 @@ declare namespace GroundStation {
     /**
      * UUID of a contact.
      */
-    contactId: String;
+    contactId: Uuid;
   }
   export type ConfigArn = string;
   export type ConfigCapabilityType = "antenna-downlink"|"antenna-downlink-demod-decode"|"antenna-uplink"|"dataflow-endpoint"|"tracking"|"uplink-echo"|"s3-recording"|string;
@@ -343,13 +385,13 @@ declare namespace GroundStation {
     /**
      * UUID of a contact.
      */
-    contactId?: String;
+    contactId?: Uuid;
     /**
      * Status of a contact.
      */
     contactStatus?: ContactStatus;
     /**
-     * End time of a contact.
+     * End time of a contact in UTC.
      */
     endTime?: Timestamp;
     /**
@@ -385,7 +427,7 @@ declare namespace GroundStation {
      */
     satelliteArn?: satelliteArn;
     /**
-     * Start time of a contact.
+     * Start time of a contact in UTC.
      */
     startTime?: Timestamp;
     /**
@@ -397,7 +439,7 @@ declare namespace GroundStation {
     /**
      * UUID of a contact.
      */
-    contactId?: String;
+    contactId?: Uuid;
   }
   export type ContactList = ContactData[];
   export type ContactStatus = "AVAILABLE"|"AWS_CANCELLED"|"AWS_FAILED"|"CANCELLED"|"CANCELLING"|"COMPLETED"|"FAILED"|"FAILED_TO_SCHEDULE"|"PASS"|"POSTPASS"|"PREPASS"|"SCHEDULED"|"SCHEDULING"|string;
@@ -425,6 +467,40 @@ declare namespace GroundStation {
      */
     tags?: TagsMap;
   }
+  export interface CreateEphemerisRequest {
+    /**
+     * Whether to set the ephemeris status to ENABLED after validation. Setting this to false will set the ephemeris status to DISABLED after validation.
+     */
+    enabled?: Boolean;
+    /**
+     * Ephemeris data.
+     */
+    ephemeris?: EphemerisData;
+    /**
+     * An overall expiration time for the ephemeris in UTC, after which it will become EXPIRED.
+     */
+    expirationTime?: Timestamp;
+    /**
+     * The ARN of a KMS key used to encrypt the ephemeris in Ground Station.
+     */
+    kmsKeyArn?: KeyArn;
+    /**
+     * A name string associated with the ephemeris. Used as a human-readable identifier for the ephemeris.
+     */
+    name: SafeName;
+    /**
+     * Customer-provided priority score to establish the order in which overlapping ephemerides should be used. The default for customer-provided ephemeris priority is 1, and higher numbers take precedence. Priority must be 1 or greater
+     */
+    priority?: CustomerEphemerisPriority;
+    /**
+     * AWS Ground Station satellite ID for this ephemeris.
+     */
+    satelliteId: Uuid;
+    /**
+     * Tags assigned to an ephemeris.
+     */
+    tags?: TagsMap;
+  }
   export interface CreateMissionProfileRequest {
     /**
      * Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.
@@ -441,7 +517,7 @@ declare namespace GroundStation {
     /**
      * Smallest amount of time in seconds that you’d like to see for an available contact. AWS Ground Station will not present you with contacts shorter than this duration.
      */
-    minimumViableContactDurationSeconds: DurationInSeconds;
+    minimumViableContactDurationSeconds: PositiveDurationInSeconds;
     /**
      * Name of a mission profile.
      */
@@ -456,6 +532,7 @@ declare namespace GroundStation {
     trackingConfigArn: ConfigArn;
   }
   export type Criticality = "PREFERRED"|"REMOVED"|"REQUIRED"|string;
+  export type CustomerEphemerisPriority = number;
   export interface DataflowDetail {
     destination?: Destination;
     /**
@@ -499,7 +576,7 @@ declare namespace GroundStation {
     /**
      * UUID of a dataflow endpoint group.
      */
-    dataflowEndpointGroupId?: String;
+    dataflowEndpointGroupId?: Uuid;
   }
   export type DataflowEndpointGroupList = DataflowEndpointListItem[];
   export interface DataflowEndpointListItem {
@@ -510,7 +587,7 @@ declare namespace GroundStation {
     /**
      * UUID of a dataflow endpoint group.
      */
-    dataflowEndpointGroupId?: String;
+    dataflowEndpointGroupId?: Uuid;
   }
   export type DataflowEndpointMtuInteger = number;
   export type DataflowList = DataflowDetail[];
@@ -524,7 +601,7 @@ declare namespace GroundStation {
     /**
      * UUID of a Config.
      */
-    configId: String;
+    configId: Uuid;
     /**
      * Type of a Config.
      */
@@ -534,13 +611,19 @@ declare namespace GroundStation {
     /**
      * UUID of a dataflow endpoint group.
      */
-    dataflowEndpointGroupId: String;
+    dataflowEndpointGroupId: Uuid;
+  }
+  export interface DeleteEphemerisRequest {
+    /**
+     * The AWS Ground Station ephemeris ID.
+     */
+    ephemerisId: Uuid;
   }
   export interface DeleteMissionProfileRequest {
     /**
      * UUID of a mission profile.
      */
-    missionProfileId: String;
+    missionProfileId: Uuid;
   }
   export interface DemodulationConfig {
     /**
@@ -552,13 +635,13 @@ declare namespace GroundStation {
     /**
      * UUID of a contact.
      */
-    contactId: String;
+    contactId: Uuid;
   }
   export interface DescribeContactResponse {
     /**
      * UUID of a contact.
      */
-    contactId?: String;
+    contactId?: Uuid;
     /**
      * Status of a contact.
      */
@@ -568,7 +651,7 @@ declare namespace GroundStation {
      */
     dataflowList?: DataflowList;
     /**
-     * End time of a contact.
+     * End time of a contact in UTC.
      */
     endTime?: Timestamp;
     /**
@@ -604,11 +687,59 @@ declare namespace GroundStation {
      */
     satelliteArn?: satelliteArn;
     /**
-     * Start time of a contact.
+     * Start time of a contact in UTC.
      */
     startTime?: Timestamp;
     /**
      * Tags assigned to a contact.
+     */
+    tags?: TagsMap;
+  }
+  export interface DescribeEphemerisRequest {
+    /**
+     * The AWS Ground Station ephemeris ID.
+     */
+    ephemerisId: Uuid;
+  }
+  export interface DescribeEphemerisResponse {
+    /**
+     * The time the ephemeris was uploaded in UTC.
+     */
+    creationTime?: Timestamp;
+    /**
+     * Whether or not the ephemeris is enabled.
+     */
+    enabled?: Boolean;
+    /**
+     * The AWS Ground Station ephemeris ID.
+     */
+    ephemerisId?: Uuid;
+    /**
+     * Reason that an ephemeris failed validation. Only provided for ephemerides with INVALID status.
+     */
+    invalidReason?: EphemerisInvalidReason;
+    /**
+     * A name string associated with the ephemeris. Used as a human-readable identifier for the ephemeris.
+     */
+    name?: SafeName;
+    /**
+     * Customer-provided priority score to establish the order in which overlapping ephemerides should be used. The default for customer-provided ephemeris priority is 1, and higher numbers take precedence. Priority must be 1 or greater
+     */
+    priority?: EphemerisPriority;
+    /**
+     * The AWS Ground Station satellite ID associated with ephemeris.
+     */
+    satelliteId?: Uuid;
+    /**
+     * The status of the ephemeris.
+     */
+    status?: EphemerisStatus;
+    /**
+     * Supplied ephemeris data.
+     */
+    suppliedData?: EphemerisTypeDescription;
+    /**
+     * Tags assigned to an ephemeris.
      */
     tags?: TagsMap;
   }
@@ -620,7 +751,7 @@ declare namespace GroundStation {
     /**
      * UUID of a Config.
      */
-    configId?: String;
+    configId?: Uuid;
     /**
      * Type of a Config.
      */
@@ -659,12 +790,90 @@ declare namespace GroundStation {
      */
     endpoint?: DataflowEndpoint;
     /**
-     * Endpoint security details.
+     * Endpoint security details including a list of subnets, a list of security groups and a role to connect streams to instances.
      */
     securityDetails?: SecurityDetails;
   }
   export type EndpointDetailsList = EndpointDetails[];
   export type EndpointStatus = "created"|"creating"|"deleted"|"deleting"|"failed"|string;
+  export type EphemeridesList = EphemerisItem[];
+  export interface EphemerisData {
+    oem?: OEMEphemeris;
+    tle?: TLEEphemeris;
+  }
+  export interface EphemerisDescription {
+    /**
+     * Supplied ephemeris data.
+     */
+    ephemerisData?: UnboundedString;
+    /**
+     * Source S3 object used for the ephemeris.
+     */
+    sourceS3Object?: S3Object;
+  }
+  export interface EphemerisIdResponse {
+    /**
+     * The AWS Ground Station ephemeris ID.
+     */
+    ephemerisId?: Uuid;
+  }
+  export type EphemerisInvalidReason = "METADATA_INVALID"|"TIME_RANGE_INVALID"|"TRAJECTORY_INVALID"|"KMS_KEY_INVALID"|"VALIDATION_ERROR"|string;
+  export interface EphemerisItem {
+    /**
+     * The time the ephemeris was uploaded in UTC.
+     */
+    creationTime?: Timestamp;
+    /**
+     * Whether or not the ephemeris is enabled.
+     */
+    enabled?: Boolean;
+    /**
+     * The AWS Ground Station ephemeris ID.
+     */
+    ephemerisId?: Uuid;
+    /**
+     * A name string associated with the ephemeris. Used as a human-readable identifier for the ephemeris.
+     */
+    name?: SafeName;
+    /**
+     * Customer-provided priority score to establish the order in which overlapping ephemerides should be used. The default for customer-provided ephemeris priority is 1, and higher numbers take precedence. Priority must be 1 or greater
+     */
+    priority?: EphemerisPriority;
+    /**
+     * Source S3 object used for the ephemeris.
+     */
+    sourceS3Object?: S3Object;
+    /**
+     * The status of the ephemeris.
+     */
+    status?: EphemerisStatus;
+  }
+  export interface EphemerisMetaData {
+    /**
+     * UUID of a customer-provided ephemeris. This field is not populated for default ephemerides from Space Track.
+     */
+    ephemerisId?: Uuid;
+    /**
+     * The epoch of a default, ephemeris from Space Track in UTC. This field is not populated for customer-provided ephemerides.
+     */
+    epoch?: Timestamp;
+    /**
+     * A name string associated with the ephemeris. Used as a human-readable identifier for the ephemeris. A name is only returned for customer-provider ephemerides that have a name associated.
+     */
+    name?: SafeName;
+    /**
+     * The EphemerisSource that generated a given ephemeris.
+     */
+    source: EphemerisSource;
+  }
+  export type EphemerisPriority = number;
+  export type EphemerisSource = "CUSTOMER_PROVIDED"|"SPACE_TRACK"|string;
+  export type EphemerisStatus = "VALIDATING"|"INVALID"|"ERROR"|"ENABLED"|"DISABLED"|"EXPIRED"|string;
+  export type EphemerisStatusList = EphemerisStatus[];
+  export interface EphemerisTypeDescription {
+    oem?: EphemerisDescription;
+    tle?: EphemerisDescription;
+  }
   export interface Frequency {
     /**
      * Frequency units.
@@ -690,7 +899,7 @@ declare namespace GroundStation {
     /**
      * UUID of a Config.
      */
-    configId: String;
+    configId: Uuid;
     /**
      * Type of a Config.
      */
@@ -726,7 +935,7 @@ declare namespace GroundStation {
     /**
      * UUID of a dataflow endpoint group.
      */
-    dataflowEndpointGroupId: String;
+    dataflowEndpointGroupId: Uuid;
   }
   export interface GetDataflowEndpointGroupResponse {
     /**
@@ -736,7 +945,7 @@ declare namespace GroundStation {
     /**
      * UUID of a dataflow endpoint group.
      */
-    dataflowEndpointGroupId?: String;
+    dataflowEndpointGroupId?: Uuid;
     /**
      * Details of a dataflow endpoint.
      */
@@ -750,11 +959,11 @@ declare namespace GroundStation {
     /**
      * The month being requested, with a value of 1-12.
      */
-    month: Integer;
+    month: Month;
     /**
      * The year being requested, in the format of YYYY.
      */
-    year: Integer;
+    year: Year;
   }
   export interface GetMinuteUsageResponse {
     /**
@@ -782,7 +991,7 @@ declare namespace GroundStation {
     /**
      * UUID of a mission profile.
      */
-    missionProfileId: String;
+    missionProfileId: Uuid;
   }
   export interface GetMissionProfileResponse {
     /**
@@ -800,7 +1009,7 @@ declare namespace GroundStation {
     /**
      * Smallest amount of time in seconds that you’d like to see for an available contact. AWS Ground Station will not present you with contacts shorter than this duration.
      */
-    minimumViableContactDurationSeconds?: DurationInSeconds;
+    minimumViableContactDurationSeconds?: PositiveDurationInSeconds;
     /**
      * ARN of a mission profile.
      */
@@ -808,15 +1017,15 @@ declare namespace GroundStation {
     /**
      * UUID of a mission profile.
      */
-    missionProfileId?: String;
+    missionProfileId?: Uuid;
     /**
      * Name of a mission profile.
      */
-    name?: String;
+    name?: SafeName;
     /**
      * Region of a mission profile.
      */
-    region?: String;
+    region?: AWSRegion;
     /**
      * Tags assigned to a mission profile.
      */
@@ -830,9 +1039,13 @@ declare namespace GroundStation {
     /**
      * UUID of a satellite.
      */
-    satelliteId: String;
+    satelliteId: Uuid;
   }
   export interface GetSatelliteResponse {
+    /**
+     * The current ephemeris being used to compute the trajectory of the satellite.
+     */
+    currentEphemeris?: EphemerisMetaData;
     /**
      * A list of ground stations to which the satellite is on-boarded.
      */
@@ -854,29 +1067,31 @@ declare namespace GroundStation {
     /**
      * UUID of a ground station.
      */
-    groundStationId?: String;
+    groundStationId?: GroundStationName;
     /**
      * Name of a ground station.
      */
-    groundStationName?: String;
+    groundStationName?: GroundStationName;
     /**
      * Ground station Region.
      */
-    region?: String;
+    region?: AWSRegion;
   }
-  export type GroundStationIdList = String[];
+  export type GroundStationIdList = GroundStationName[];
   export type GroundStationList = GroundStationData[];
+  export type GroundStationName = string;
   export type Integer = number;
   export type JsonString = string;
+  export type KeyArn = string;
   export interface ListConfigsRequest {
     /**
      * Maximum number of Configs returned.
      */
-    maxResults?: Integer;
+    maxResults?: PaginationMaxResults;
     /**
      * Next token returned in the request of a previous ListConfigs call. Used to get the next page of results.
      */
-    nextToken?: String;
+    nextToken?: PaginationToken;
   }
   export interface ListConfigsResponse {
     /**
@@ -886,21 +1101,21 @@ declare namespace GroundStation {
     /**
      * Next token returned in the response of a previous ListConfigs call. Used to get the next page of results.
      */
-    nextToken?: String;
+    nextToken?: PaginationToken;
   }
   export interface ListContactsRequest {
     /**
-     * End time of a contact.
+     * End time of a contact in UTC.
      */
     endTime: Timestamp;
     /**
      * Name of a ground station.
      */
-    groundStation?: String;
+    groundStation?: GroundStationName;
     /**
      * Maximum number of contacts returned.
      */
-    maxResults?: Integer;
+    maxResults?: PaginationMaxResults;
     /**
      * ARN of a mission profile.
      */
@@ -908,13 +1123,13 @@ declare namespace GroundStation {
     /**
      * Next token returned in the request of a previous ListContacts call. Used to get the next page of results.
      */
-    nextToken?: String;
+    nextToken?: PaginationToken;
     /**
      * ARN of a satellite.
      */
     satelliteArn?: satelliteArn;
     /**
-     * Start time of a contact.
+     * Start time of a contact in UTC.
      */
     startTime: Timestamp;
     /**
@@ -930,17 +1145,17 @@ declare namespace GroundStation {
     /**
      * Next token returned in the response of a previous ListContacts call. Used to get the next page of results.
      */
-    nextToken?: String;
+    nextToken?: PaginationToken;
   }
   export interface ListDataflowEndpointGroupsRequest {
     /**
      * Maximum number of dataflow endpoint groups returned.
      */
-    maxResults?: Integer;
+    maxResults?: PaginationMaxResults;
     /**
      * Next token returned in the request of a previous ListDataflowEndpointGroups call. Used to get the next page of results.
      */
-    nextToken?: String;
+    nextToken?: PaginationToken;
   }
   export interface ListDataflowEndpointGroupsResponse {
     /**
@@ -950,21 +1165,57 @@ declare namespace GroundStation {
     /**
      * Next token returned in the response of a previous ListDataflowEndpointGroups call. Used to get the next page of results.
      */
-    nextToken?: String;
+    nextToken?: PaginationToken;
+  }
+  export interface ListEphemeridesRequest {
+    /**
+     * The end time to list in UTC. The operation will return an ephemeris if its expiration time is within the time range defined by the startTime and endTime.
+     */
+    endTime: Timestamp;
+    /**
+     * Maximum number of ephemerides to return.
+     */
+    maxResults?: PaginationMaxResults;
+    /**
+     * Pagination token.
+     */
+    nextToken?: PaginationToken;
+    /**
+     * The AWS Ground Station satellite ID to list ephemeris for.
+     */
+    satelliteId: Uuid;
+    /**
+     * The start time to list in UTC. The operation will return an ephemeris if its expiration time is within the time range defined by the startTime and endTime.
+     */
+    startTime: Timestamp;
+    /**
+     * The list of ephemeris status to return.
+     */
+    statusList?: EphemerisStatusList;
+  }
+  export interface ListEphemeridesResponse {
+    /**
+     * List of ephemerides.
+     */
+    ephemerides?: EphemeridesList;
+    /**
+     * Pagination token.
+     */
+    nextToken?: PaginationToken;
   }
   export interface ListGroundStationsRequest {
     /**
      * Maximum number of ground stations returned.
      */
-    maxResults?: Integer;
+    maxResults?: PaginationMaxResults;
     /**
      * Next token that can be supplied in the next call to get the next page of ground stations.
      */
-    nextToken?: String;
+    nextToken?: PaginationToken;
     /**
      * Satellite ID to retrieve on-boarded ground stations.
      */
-    satelliteId?: String;
+    satelliteId?: Uuid;
   }
   export interface ListGroundStationsResponse {
     /**
@@ -974,17 +1225,17 @@ declare namespace GroundStation {
     /**
      * Next token that can be supplied in the next call to get the next page of ground stations.
      */
-    nextToken?: String;
+    nextToken?: PaginationToken;
   }
   export interface ListMissionProfilesRequest {
     /**
      * Maximum number of mission profiles returned.
      */
-    maxResults?: Integer;
+    maxResults?: PaginationMaxResults;
     /**
      * Next token returned in the request of a previous ListMissionProfiles call. Used to get the next page of results.
      */
-    nextToken?: String;
+    nextToken?: PaginationToken;
   }
   export interface ListMissionProfilesResponse {
     /**
@@ -994,23 +1245,23 @@ declare namespace GroundStation {
     /**
      * Next token returned in the response of a previous ListMissionProfiles call. Used to get the next page of results.
      */
-    nextToken?: String;
+    nextToken?: PaginationToken;
   }
   export interface ListSatellitesRequest {
     /**
      * Maximum number of satellites returned.
      */
-    maxResults?: Integer;
+    maxResults?: PaginationMaxResults;
     /**
      * Next token that can be supplied in the next call to get the next page of satellites.
      */
-    nextToken?: String;
+    nextToken?: PaginationToken;
   }
   export interface ListSatellitesResponse {
     /**
      * Next token that can be supplied in the next call to get the next page of satellites.
      */
-    nextToken?: String;
+    nextToken?: PaginationToken;
     /**
      * List of satellites.
      */
@@ -1020,7 +1271,7 @@ declare namespace GroundStation {
     /**
      * ARN of a resource.
      */
-    resourceArn: String;
+    resourceArn: AnyArn;
   }
   export interface ListTagsForResourceResponse {
     /**
@@ -1033,7 +1284,7 @@ declare namespace GroundStation {
     /**
      * UUID of a mission profile.
      */
-    missionProfileId?: String;
+    missionProfileId?: Uuid;
   }
   export type MissionProfileList = MissionProfileListItem[];
   export interface MissionProfileListItem {
@@ -1044,26 +1295,40 @@ declare namespace GroundStation {
     /**
      * UUID of a mission profile.
      */
-    missionProfileId?: String;
+    missionProfileId?: Uuid;
     /**
      * Name of a mission profile.
      */
-    name?: String;
+    name?: SafeName;
     /**
      * Region of a mission profile.
      */
-    region?: String;
+    region?: AWSRegion;
   }
+  export type Month = number;
+  export interface OEMEphemeris {
+    /**
+     * The data for an OEM ephemeris, supplied directly in the request rather than through an S3 object.
+     */
+    oemData?: UnboundedString;
+    /**
+     * Identifies the S3 object to be used as the ephemeris.
+     */
+    s3Object?: S3Object;
+  }
+  export type PaginationMaxResults = number;
+  export type PaginationToken = string;
   export type Polarization = "LEFT_HAND"|"NONE"|"RIGHT_HAND"|string;
+  export type PositiveDurationInSeconds = number;
   export interface ReserveContactRequest {
     /**
-     * End time of a contact.
+     * End time of a contact in UTC.
      */
     endTime: Timestamp;
     /**
      * Name of a ground station.
      */
-    groundStation: String;
+    groundStation: GroundStationName;
     /**
      * ARN of a mission profile.
      */
@@ -1073,7 +1338,7 @@ declare namespace GroundStation {
      */
     satelliteArn: satelliteArn;
     /**
-     * Start time of a contact.
+     * Start time of a contact in UTC.
      */
     startTime: Timestamp;
     /**
@@ -1082,7 +1347,23 @@ declare namespace GroundStation {
     tags?: TagsMap;
   }
   export type RoleArn = string;
+  export type S3BucketName = string;
   export type S3KeyPrefix = string;
+  export interface S3Object {
+    /**
+     * An Amazon S3 Bucket name.
+     */
+    bucket?: S3BucketName;
+    /**
+     * An Amazon S3 key for the ephemeris.
+     */
+    key?: S3ObjectKey;
+    /**
+     * For versioned S3 objects, the version to use for the ephemeris.
+     */
+    version?: S3VersionId;
+  }
+  export type S3ObjectKey = string;
   export interface S3RecordingConfig {
     /**
      * ARN of the bucket to record to.
@@ -1103,13 +1384,18 @@ declare namespace GroundStation {
      */
     bucketArn?: BucketArn;
     /**
-     * Template of the S3 key used.
+     * Key template used for the S3 Recording Configuration
      */
     keyTemplate?: String;
   }
+  export type S3VersionId = string;
   export type SafeName = string;
   export type SatelliteList = SatelliteListItem[];
   export interface SatelliteListItem {
+    /**
+     * The current ephemeris being used to compute the trajectory of the satellite.
+     */
+    currentEphemeris?: EphemerisMetaData;
     /**
      * A list of ground stations to which the satellite is on-boarded.
      */
@@ -1154,7 +1440,7 @@ declare namespace GroundStation {
   }
   export interface Source {
     /**
-     * Additional details for a Config, if type is dataflow endpoint or antenna demod decode.
+     * Additional details for a Config, if type is dataflow-endpoint or antenna-downlink-demod-decode 
      */
     configDetails?: ConfigDetails;
     /**
@@ -1187,12 +1473,37 @@ declare namespace GroundStation {
   export type StatusList = ContactStatus[];
   export type String = string;
   export type SubnetList = String[];
-  export type TagKeys = String[];
+  export interface TLEData {
+    /**
+     * First line of two-line element set (TLE) data.
+     */
+    tleLine1: TleLineOne;
+    /**
+     * Second line of two-line element set (TLE) data.
+     */
+    tleLine2: TleLineTwo;
+    /**
+     * The valid time range for the TLE. Gaps or overlap are not permitted.
+     */
+    validTimeRange: TimeRange;
+  }
+  export type TLEDataList = TLEData[];
+  export interface TLEEphemeris {
+    /**
+     * Identifies the S3 object to be used as the ephemeris.
+     */
+    s3Object?: S3Object;
+    /**
+     * The data for a TLE ephemeris, supplied directly in the request rather than through an S3 object.
+     */
+    tleData?: TLEDataList;
+  }
+  export type TagKeys = UnboundedString[];
   export interface TagResourceRequest {
     /**
      * ARN of a resource tag.
      */
-    resourceArn: String;
+    resourceArn: AnyArn;
     /**
      * Tags assigned to a resource.
      */
@@ -1201,18 +1512,31 @@ declare namespace GroundStation {
   export interface TagResourceResponse {
   }
   export type TagsMap = {[key: string]: String};
+  export interface TimeRange {
+    /**
+     * Time in UTC at which the time range ends.
+     */
+    endTime: Timestamp;
+    /**
+     * Time in UTC at which the time range starts.
+     */
+    startTime: Timestamp;
+  }
   export type Timestamp = Date;
+  export type TleLineOne = string;
+  export type TleLineTwo = string;
   export interface TrackingConfig {
     /**
      * Current setting for autotrack.
      */
     autotrack: Criticality;
   }
+  export type UnboundedString = string;
   export interface UntagResourceRequest {
     /**
      * ARN of a resource.
      */
-    resourceArn: String;
+    resourceArn: AnyArn;
     /**
      * Keys of a resource tag.
      */
@@ -1228,7 +1552,7 @@ declare namespace GroundStation {
     /**
      * UUID of a Config.
      */
-    configId: String;
+    configId: Uuid;
     /**
      * Type of a Config.
      */
@@ -1237,6 +1561,24 @@ declare namespace GroundStation {
      * Name of a Config.
      */
     name: SafeName;
+  }
+  export interface UpdateEphemerisRequest {
+    /**
+     * Whether the ephemeris is enabled or not. Changing this value will not require the ephemeris to be re-validated.
+     */
+    enabled: Boolean;
+    /**
+     * The AWS Ground Station ephemeris ID.
+     */
+    ephemerisId: Uuid;
+    /**
+     * A name string associated with the ephemeris. Used as a human-readable identifier for the ephemeris.
+     */
+    name?: SafeName;
+    /**
+     * Customer-provided priority score to establish the order in which overlapping ephemerides should be used. The default for customer-provided ephemeris priority is 1, and higher numbers take precedence. Priority must be 1 or greater
+     */
+    priority?: EphemerisPriority;
   }
   export interface UpdateMissionProfileRequest {
     /**
@@ -1254,11 +1596,11 @@ declare namespace GroundStation {
     /**
      * Smallest amount of time in seconds that you’d like to see for an available contact. AWS Ground Station will not present you with contacts shorter than this duration.
      */
-    minimumViableContactDurationSeconds?: DurationInSeconds;
+    minimumViableContactDurationSeconds?: PositiveDurationInSeconds;
     /**
      * UUID of a mission profile.
      */
-    missionProfileId: String;
+    missionProfileId: Uuid;
     /**
      * Name of a mission profile.
      */
@@ -1289,6 +1631,7 @@ declare namespace GroundStation {
     polarization?: Polarization;
   }
   export type Uuid = string;
+  export type Year = number;
   export type noradSatelliteID = number;
   export type satelliteArn = string;
   /**