+++ clients/mediapackagevod.d.ts
@@ -396,6 +396,10 @@ rounded to the nearest multiple of the source fragment duration.
      */
     IncludeEncoderConfigurationInSegments?: __boolean;
     /**
+     * When enabled, an I-Frame only stream will be included in the output.
+     */
+    IncludeIframeOnlyStream?: __boolean;
+    /**
      * A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH)
 Media Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not
 be partitioned into more than one period. If the list contains "ADS", new periods will be created where